fixup! JAMES-3096 filter properties depending on client provided capabilities
